+++ package.json
@@ -1,6 +1,7 @@
 {
 	"name": "vroom",
 	"version": "0.0.1",
+	"private": true,
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
@@ -16,44 +17,39 @@
 		"migrate": "drizzle-kit migrate"
 	},
 	"devDependencies": {
-		"@fontsource/fira-mono": "^4.5.10",
-		"@neoconfetti/svelte": "^1.0.0",
 		"@playwright/test": "^1.28.1",
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/adapter-node": "^5.2.0",
 		"@sveltejs/kit": "^2.0.0",
 		"@sveltejs/vite-plugin-svelte": "^3.0.0",
 		"@tsconfig/svelte": "^5.0.4",
-		"@types/eslint": "^8.56.7",
-		"@types/node": "^20.14.9",
+		"@types/eslint": "^9.6.0",
 		"@types/pg": "^8.11.6",
 		"@types/uuid": "^10.0.0",
 		"autoprefixer": "^10.4.19",
-		"drizzle-kit": "^0.22.7",
+		"drizzle-kit": "^0.23.0",
 		"eslint": "^9.0.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0",
 		"globals": "^15.0.0",
-		"postcss": "^8.4.39",
+		"postcss": "^8.4.40",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
 		"svelte": "^4.2.7",
 		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.4.4",
-		"tslib": "^2.4.1",
-		"typescript": "^5.5.3",
+		"tailwindcss": "^3.4.7",
+		"tslib": "^2.6.3",
+		"typescript": "^5.0.0",
 		"typescript-eslint": "^8.0.0-alpha.20",
-		"vite": "^5.3.3",
-		"vitest": "^1.2.0"
+		"vite": "^5.0.3",
+		"vitest": "^2.0.0"
 	},
 	"type": "module",
 	"dependencies": {
 		"@lucia-auth/adapter-drizzle": "^1.0.7",
-		"dotenv": "^16.4.5",
-		"drizzle-orm": "^0.31.2",
-		"lucia": "^3.2.0",
-		"pg": "^8.12.0",
+		"drizzle-orm": "^0.32.1",
 		"fast-glob": "^3.3.2",
-		"mapbox-gl": "^3.5.0"
+		"lucia": "^3.2.0",
+		"mapbox-gl": "^3.5.2",
+		"pg": "^8.12.0"
 	}
 }
