+++ src/app.d.ts
@@ -9,6 +9,7 @@ declare global {
 			user: User & { balance: string | null, isAdmin: boolean | null, role: "customer" | "employee" | "admin" } | null;
 			session: Session | null;
 		}
+		// interface PageData {}
 		// interface PageState {}
 		// interface Platform {}
 	}
