+++ src/lib/components/MapComponent.svelte
