+++ src/lib/components/Sidebar.svelte
