+++ src/routes/(authed)/+layout.svelte
@@ -1,11 +1,13 @@
 <script lang="ts">
 	import { enhance } from '$app/forms';
+	import type { LayoutData } from './$types';
 	
-	export let data;
+	export let data: LayoutData;
+	const { user } = data;
 </script>
 
 <main>
-	<h1>Hello {data.firstname}!</h1>
+	<h1>Hello {user.firstname}!</h1>
 	<form method="post" action="/logout" use:enhance>
 		<button>Sign out</button>
 	</form>
