--- src/routes/+page.ts
@@ -1,2 +1,0 @@
-// We use Actions and dynamic content so this page cannot be pre-rendered
-export const prerender = false;
